+++ SourceCode/App/Components/AppInputBase.razor.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Localization;
 using ModulesRegistry.Data.Extensions;
+using ModulesRegistry.Extensions;
 using System.Linq.Expressions;
 
 namespace ModulesRegistry.Components;
@@ -20,10 +21,9 @@ public abstract partial class AppInputBase<TValue>
 
     public string WidthCss => Width.HasValue ? $"col-md-{Width}" : "col-md-4";
 
-    public string LabelText(IStringLocalizer<App> localizer) 
+    public string LabelText(IStringLocalizer<App> localizer)
     {
         if (Label.HasNoValue()) return string.Empty;
-        var items = Label.Split('-');
-        return string.Join('/', items.Select(x => localizer[x]));
+        return localizer.FromParts(Label);
     }
 }