+++ SourceCode/Data/EnginePowerUnit.cs
@@ -1,0 +1,10 @@
+﻿#nullable disable
+
+namespace ModulesRegistry.Data;
+
+public enum EnginePowerUnit
+{
+    Undefined = 0,
+    Kilowatts = 1,
+    Horsepower = 2,
+}