+++ SourceCode/Data/Extensions/StringExtensions.cs
@@ -27,6 +27,10 @@ public static class StringExtensions
         string.IsNullOrWhiteSpace(value) ? string.Empty :
         value.Length == 1 ? value.ToLowerInvariant() :
         string.Concat(value[0].ToString().ToLowerInvariant(), value.AsSpan(1));
+    public static string ToFirstUpperInvariant(this string? value) =>
+         string.IsNullOrWhiteSpace(value) ? string.Empty :
+         value.Length == 1 ? value.ToUpperInvariant() :
+         string.Concat(value[0].ToString().ToUpperInvariant(), value.AsSpan(1));
 
     public static string[] Items(this string? value, char separator = ';') =>
         string.IsNullOrWhiteSpace(value) ? [] :